clean up the mess relating to the "Insufficient free cell and/or spaces" message
It was hardcoded in freecell.js, and could turn up multiple times during a smartMove() execution.  The message is now in the dtd file, and retrieved through an attr on the window element.  canMoveTo now returns 0 when there are insufficient spaces, which is ==false, but !==false, allowing smartMove to skip the alert()s, but our custom attemptMove() to show them