various fixes and two new games
Grandfather:
renamed to Russian Solitaire, the name it has in PySol

Spider:
oddities of canMoveTo, moveTo etc (caused by the moves being to the foundationsBox rather than a specific foundation) mostly eliminated by having moveTo just pick the first empty foundation when the destination for the move is any of the foundations.

New games:
Black Widow - a spider variant. code is in spider.js
Sanibel - a game very like yukon, except for the card layout

General:
Card game names, accesskeys for the menu, and difficuly level descriptions all moved from the dtd file to a stringbundle, so that different games (Black Widow and Spider) can share the same XUL.  Difficulty levels are now specified using the |difficultyLevels| field of a game, rather than with an attribute in the XUL file.