code cleanup
Change the undo/move-tracking system to use objects that have perform() and undo() methods, so that different types of move are all undone in the same way.  (Should allow for Redo too.)  This includes eliminating BaseCardGame.trackMove() in favour of the new doAction()

Use rule_dealFromStock to handle the different systems for how the stock should work (dealing to the waste pile, to the waste pile and allowing the stock to be turned over, and dealing to each of stacks[]). rather than having a method in BaseCardGame that tries to handle all cases.  stockDealTargets[] and canTurnStockOver are therefore eliminated.

Remove most of the redealing code in BaseCardGame, since nothing uses it any more.

Scoring in Mod3 (and Golf?) is broken by these changes.  Will fix later.