make animation fast again!!!
This margin+padding hackery was to prevent the card piles spreading out to the right if a card was dragged beyond the right hand edge of the window, but it hasn't been achieving that for a while.  It was also to prevent the animation and/or dragging from being very slow if some card were lower than the bottom of the window, but overflow:hidden takes care of that.