Vastly simplify all the code for computing hints
The old hint code was quite complex due to trying to prioritise hints (showing moves to empty piles last, for example).  This is much less important now multiple hint destinations can be shown at the same time, so a very simple hint-calculation alogorithm now suffices.

For games like Golf it would probably be better to highlight all the movable cards (since they are all destined for the foundation), but this patch does not do that.