Eliminate Game.done().  Move code for turning revealed cards face up into Game.doo()
Formerly, Game.done() had to be called after animations finished because only then would the piles be in the right states for it, but that is no longer the case.  This change is another step to making animations a purely decorative thing.

A minor side-effect of this patch is to immediately reveal the face up newly-exposed cards (rather than doing so only after the causing move's animation completes).  While different, this looks fine.