fix pyramid and tripeaks layout.
They were broken by setting .style.MozBoxAlign="start" on hboxes, which historically was suppressed by {align=} in the .xulTemplate.  The workaround doesn't work now we're using HTML since @align isn't special, but the MozBoxAlign thing can now be done in CSS, and overridden there.  This makes the {} notation in templates needlessly complex, so change it to just be a way of adding a className value.  And make it *add* the class, rather than overwriting, which didn't matter in XUL, but does now we're encoding hbox vs vbox in className rather than tagName.