make the mouse work properly in Pyramid and TriPeaks.
Prior to this (and dating back a while, certainly before the display:-moz-box -> <table> conversion) clicks would only work in the upper portion of the card, i.e. that not covered by the next row down, even when the next tow was supposed to be hidden).  The new code makes all the <table>s for each row 0-height, with position:relative <td>s and position:absolute <canvas>es hanging out of them.