introduce a new game I invented called "Double Simon", like Simple Simon but with two decks of cards and a different layout
I gave it 12 piles (rather than 10 in Simple Simon) because obviously it needs a bit more space, and 12 still fits on my 1280x800 screen.  And I gave it two piles with just 1 card because you really need to get spaces for it to work, and that gives more opportunity.

So far it mostly seems to be winnable, using undo a lot, and taking >250 moves.