change .deal() to use a new ._deal_cards() that's simpler than the old ._dealSomeCards()
The .splice() stuff was causing weirdness in a Double Fan game I'm in the process of writing (the last four cards would get 1,2,1,0 across four spaces, rather than all in one).  This version seems simple overall, and fixes that.

Changing Penguin is harder, and I'm considering rewriting it anyway, so let it just keep ._dealSomeCards() for now.

Double Simon had a bogus count in its .pileDetails, which happened to work with the old code but no longer worked.