fix calculation of offsets for _FlexFanView with only one card
Previously the code always calculated an offset of zero.  That was fine for displaying the card itself, but meant that when the pile was the destination for a hint, the ghosted cards would be drawn without an appropriate offset from those already present.

The fix is to eliminate the broken "if(num <= 0) return", and handle 0 properly instead.  Also, clean things up by pushing the num-1 from the caller to the callee, and de-duplicate the horizontal vs vertical calculations.