fix pyramid/tripeaks so the piles are visible
Basically, to get the overlapping piles while still having mouse events flow to the correct pile view, we make all the layout elements have zero height (and margins for row-spacing), and then position:absolute the <canvas>es.  So you'd expect that to require position:relative on the <td>s holding them, but actually that doesn't really work right in Firefox in general, and in this particular case it makes the <canvas>es not display at all (though I have no idea why).  Removing position:relative fixes that, and because we're not actually setting top:0 and left:0 (we're leaving them unset), the <canvas> displays in the correct place anyway.

This seems to have been broken a long time.