make hint-destination piles' views draw the ghosted cards themselves
Previously we had the source pile draw them in to gFloatingPile, and then the destination pile's view just composited that onto itself.  Now we just tell the destination pile view what extra cards to draw and let it deal with it.

In the short term it's slightly more code.  But it's conceptually cleaner.  And aesthetically better (because for fan views we now use the offset of the destination pile rather than the source pile).  And it isolates the complexity of the former .highlightHintTo() in _FanView.  And hopefully will allow re-thinking all the complex mess surrounding ._getHighlightBounds() and .getVisibleCardIndexes() etc.