simplify offset-calculation and drawing of flexible fans
The Layout was and is what really determines how much space is available for the fans, and sets it via .fixedWidth/fixedHeight (formerly via .widthToUse/heightToUse).  So we now do the simple and correct thing, and just rely on those numbers to work out the offsets.

Prior to the recent introduction of .draw_into(), ._updateOffsets() would re-measure the <canvas> using .getBoundingClientRect() rather than just using the saved dimensions.  This happened to work OK because we first resized the <canvas> to the dimensions the Layout calculated (which led to the necessity of an extra clear_and_resize_canvas() after the recent addition of .draw_into()).  This measuring from within the view is probably leftover from the XUL era, when the view's size was controlled via XUL flex="1".