rewrite autoplay_default() to be much simpler code
The new approach (copied from Gypsy) is to just try the top card of each pile, and move it if we can.  It's simple and obvious, and a quick benchmark showed it takes ~1ms.

The old approach was much more complex: for each foundations, identify the next card that could go on it, then check if it's currently moveable.  This may have made sense back when cards were XUL elements (so property accesses were very expensive, and minimising comparisons really mattered).  Also, iirc it predates .autoplayable_numbers() - that calculation was instead done as we looked at each foundation, and so iterating foundations rather than piles probably avoided O(M x N) behaviour there.