fix TriPeaks's slightly-broken .best_action_for()
It was checking card.faceUp rather than calling .mayTakeCard(), which has been incorrect since I changed TriPeaks's card to only be rendered face down rather than actually having .faceUp set to false.  This allowed you to move a card that had other cards on top of it, if you happened to know it's number was +1/-1 from the foundation card (or just clicked it at random).