store cards in Stock piles face-up, for convenience
Canfield's .deal() needs modifying because .deal_cards() no longer treats the "52" as just "all remaining cards".

TriPeak's .deal() could be modified in the same way, but it's simpler to just delete it, as it's just left over from when cards in the peaks were actually face-down on most of the piles, rather than merely being rendered as such.