reimplement most Action classes in terms of a new GenericAction
With Card now immutable, most actions can be handled easily as just a list of piles and what their .cards contained before and after the action.

Move's animation does change as a result of this: turning a card from the source pile face up now visibly happens at the end of the animate move, whereas previously it happened at the start.  I don't think this change matters (it's arguably an improvement).