dondorf.svg: replace with a new version generated by a python script
It looks almost identical, but is structured competely differently, as
it now uses a translated <g> for each card, with all the numbers etc
using coords relative to the card.

The only visible changes are the removal of the two unused blank cards,
and the pip in the middle of the 9 of clubs being in the regular place,
whereas previously it was slightly higher than the others.  I think it
looks better this way.